# Conflicts:
#	src/main/java/io/antmedia/rest/RestServiceBase.java
#	src/test/java/io/antmedia/test/security/TokenFilterTest.java

+++ src/main/java/io/antmedia/servlet/ChunkedTransferServlet.java
@@ -12,15 +12,15 @@ import java.nio.file.StandardCopyOption;
 import java.util.Collection;
 import java.util.concurrent.LinkedBlockingQueue;
 
-import javax.servlet.AsyncContext;
-import javax.servlet.AsyncEvent;
-import javax.servlet.AsyncListener;
-import javax.servlet.ServletException;
-import javax.servlet.ServletOutputStream;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.AsyncContext;
+import jakarta.servlet.AsyncEvent;
+import jakarta.servlet.AsyncListener;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletOutputStream;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 import org.apache.catalina.connector.ClientAbortException;
 import org.apache.commons.lang3.exception.ExceptionUtils;
@@ -227,8 +227,13 @@ public class ChunkedTransferServlet extends HttpServlet {
 				}
 			}
 			
-			Files.move(tmpFile.toPath(), finalFile.toPath(), StandardCopyOption.ATOMIC_MOVE, StandardCopyOption.REPLACE_EXISTING);
-			logger.info("File:{} was generated ", finalFile.getName());
+			if (finalFile.getParentFile().exists()) {
+				Files.move(tmpFile.toPath(), finalFile.toPath(), StandardCopyOption.ATOMIC_MOVE, StandardCopyOption.REPLACE_EXISTING);
+				logger.debug("File:{} was generated ", finalFile.getName());
+			}
+			else {
+				logger.warn("Parent file of {} not exist", finalFile.getParent());
+			}
 		}
 		catch (ClientAbortException e) {
 			logger.warn("Client aborted - Reading input stream for file: {}", finalFile.getAbsolutePath());
